Refactor: Move benchmarking sources to test/bench/
